requirements.txt
pandas
numpy
python-dotenv
requests
transformers
torch
scikit-learn